+++ lib/kataho_code.dart
@@ -14,4 +14,5 @@ export 'src/models/geocode_number_suffix.dart';
 export 'src/models/geocode_word.dart';
 export 'src/models/geocode_word_kid.dart';
 export 'src/models/kataho_code.dart';
+export 'src/models/kataho_code_type.dart';
 export 'src/presentation/kataho_code_builder.dart';